requirements.txt
discord.py
yt_dlp
PyNaCl
redis